requirements.txt
alembic==1.0.0
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.0.3
coveralls==1.3.0
docopt==0.6.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.12.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
psycopg2==2.7.5
py==1.5.4
pycparser==2.18
PyJWT==1.4.2
pylint==2.1.1
pytest==3.7.1
pytest-cov==2.5.1
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
Flask-Cors==3.0.6